Create Class for FirestoreMember.
Export FirestoreMember class from @community/data.
Move the declaration of  ProgressData, ProfileData, and AchievementData,
to @community/data to make them accessible to FirestoreMember class.
Create fromJSON equivalents for FirestoreMemberConverter in these classes.
Reorder constructor arguments of these classes to be in alphabetical order.
Update the createFirestoreMember cloud function to have necessary properties.

Create FirestoreConverter for FirestoreMember class.
Export memberSnap function (that fetches a firestore member) from @community/data.
Use memberSnap everywhere firestoreMember is fetched because the snap takes
the converter into account when fetching the firestoreMember.

Make progressData and achievementData to save with arrayUnion instead of as new arrays.
Did this because saving new arrays delete old progress/achievement values.

Update the isNewMember check for App and Welcome components in community.
Make it to check if the properties of the firestoreMember's profile object
are all empty strings, instead of checking if the profile object is absent.
(had to do this because all firestore users now have a profile object,
whether they had set their profile or not, courtesy of converter).